Move chain to common
